Lengte aangepast naar een float

+++ car.cpp
@@ -1,4 +1,4 @@
 #include "car.h"
 
-Car::Car(int weight, int length, int maxspeed, string name) : Vehicle(weight, length, maxspeed, name) {};
+Car::Car(int weight, float length, int maxspeed, string name) : Vehicle(weight, length, maxspeed, name) {};
 Car::~Car() {};
